// app/requests/_personal/_moderation/getPlaceModeration/getPlaceModerationSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetPlaceModerationSuccess {
  data: {
    id: number;
    name: string;
    latitude: string;
    longitude: string;
    address_kladr: string;
    logo: string | null;
    region: {
      id: number;
      name: string;
    };
    brand: {
      id: number;
      name: string;
      logo: string | null;
      description: string;
    };
  }[];
}